Second code commit, after a week's mod
Now signin.py can send logs to trigger, so we can know the status of the
program from the seewo TV.

Current improvement plans:
1. Make the socket communication better, so that even if trigger is ter-
minated accidentally, auto-DK continues.

2. Make more realistic DK times. The current all-random generator may
generate illogical information. Posting StudentNew every minute will help
construct a timeline of who entered or left when, but that wouldn't
guarantee the correctness of the data. So this is left till later.