Wrote report_absent
It relies on the fact that when a DK is on and someone haven't signed in,
his OptimisticLockField will be 1.
This program reads the database for those people's card numbers and prints
them to the screen, one each line.
This can be parsed with subprocess.run(text=True, capture_output=True)