Added GetRelay and DenyAllElse
GetRelay relays the GET request to the school server.
DenyAllElse is used to prevent malicious traffic that may cause the
network to be saturated. By design, this HTTP server should only be
accessible from the loopback interface.

When a request that's not from 127.0.0.1 is received, it is immediately
given a 403 forbidden.