Added quit_spirit and flush_notice
The first command is used for special occasions. For example, the Friday
before the Entrance Exam was on schedule, but human intelligence tells us
that no sign in was necessary. We can use this command to manually disable
the system.

The second is used for a workaround for broken GS programs, as is the
case with our machine this April.