Important bug fix: SIGSEGV in watchdog
The implementation before this commit leaves a std::promise reference
dangling after get_stu_new() returns with a time out. The detached thread
will try to set a value in the promise, which potentially writes to in-
valid memory.

After some tests, I discovered that the fourth thread created for
execute_request() will trigger this bug.

This has been resolved by putting the promise object on stack and mana-
ging the reference count with a shared pointer.