Fixed the lock db issue (observed 8.24)
When a DK session ends, the GS will lock the database and read the info
inside. (Why?)
Prior to this commit, the near_exits() call was not enclosed in a try-
catch block, leading to an uncaptured SQLError if the database is locked.
Now we put it into a try catch block that should handle the problem.
Still to be tested!