Added simul_limit >= local_limit to constraints.
Now this is checked by the `validate()` function, so the user gets an
error dialog box if this is not satisfied.

Also removed the second error dialog box saying "error with the configu-
ration file", which doesn't say much and only annoys users.

Replaced decltype(config) with Configuration along the way, since that's
simpler to type.