RAII approach to closing transactions.
If someone uses his/her card to sign in when we are trying to sign in,
a "database locked" error will be triggered in write_record(), which'll
throw an exception and skip the last line of ending the transaction.

Then all the following attempts to create new transactions in write_record()
will all fail because we can not begin another transaction with one in
place, and the transaction also effectively deprives GS of its access
to the DB.

So I used a RAII pattern to fix the bug.