Made the transactions exclusive
Before this, we often see a bug that happens at a very low probability.

When we start a deferred transaction (the default), the lock isn't obtained
on the database. So it's possible for the first write command to return
SQLITE_BUSY, which we ignored. Then the RAII guard cleans up by calling
"commit", while the doc says that you should rollback. So this causes
some sort of undefined behaviour.

Now we use an exclusive transaction, so the lock is obtained immediately
and this also prevents GS from reading intermediate states of the db.